Split layers into layers module
